+++ app/controllers/application.js
@@ -1,11 +1,15 @@
 import Controller from '@ember/controller';
 import { action } from '@ember/object';
+import { tracked } from '@glimmer/tracking';
 import { inject as service } from '@ember/service';
 
 class ControllerApplication extends Controller {
   @service session;
   @service currentUser;
   @service googleVision;
+
+  @tracked
+  textAnalyzed = 'Nothing analyzed yet';
 
   @action
   login() {
@@ -18,8 +22,19 @@ class ControllerApplication extends Controller {
   }
 
   @action
-  showSessionInfo() {
-    this.googleVision.analyzeFile('gs://redciudadana-datatesting/visionAPI/0001.jpg');
+  getTextFromFile() {
+    this.googleVision.analyzeFileFromCloudStorage('gs://redciudadana-datatesting/visionAPI/0001.jpg').then((text) => {
+      this.textAnalyzed = text;
+    });
+  }
+
+  @action
+  uploadFileAndGetText(file) {
+    file.readAsDataURL().then((fileBase64URL) => {
+      this.googleVision.analyzeAttachFile(fileBase64URL).then((text) => {
+        this.textAnalyzed = text;
+      });
+    });
   }
 }