+++ app/services/google-vision.js
@@ -7,12 +7,12 @@ import fetch from 'fetch';
 /**
  * Service to make request to Google Vision API.
  *
- * @class Navbar
+ * @class GoogleVisionService
  * @namespace Services
  * @extends Ember.Service
  * @public
- */
-class GoogleVisionService extends Service {
+*/
+export default class GoogleVisionService extends Service {
   @service session;
 
   init() {
@@ -20,7 +20,59 @@ class GoogleVisionService extends Service {
     debug('[google-vision:service] initalized!');
   }
 
-  analyzeFile(gscloudfile) {
+  /**
+   * Extract text from uploaded file encoded base64 and return a array with array per document with all text per pages.
+   * Google Vision Endpoint: `https://vision.googleapis.com/v1/files:annotate`
+   *
+   * @param {string} file The file to upload encoded in URL format
+   */
+  analyzeAttachFile(file) {
+    let accessToken = this.get('session.data.authenticated.access_token');
+
+    let body = {
+      requests: [
+        {
+          inputConfig: {
+            content: file.replace(/^data:.+;base64,/, ''),
+            mimeType: 'application/pdf'
+          },
+          features: [
+            {
+              type: 'DOCUMENT_TEXT_DETECTION'
+            }
+          ]
+        }
+      ]
+    };
+
+    return fetch('https://cors-anywhere.herokuapp.com/https://vision.googleapis.com/v1/files:annotate?alt=json', {
+      headers: {
+        accept: '*/*',
+        authorization: `Bearer ${accessToken}`,
+        'content-type': 'application/json'
+      },
+      body: JSON.stringify(body),
+      method: 'POST',
+      credentials: 'include'
+      // referrerPolicy: 'no-referrer-when-downgrade'
+    })
+    .then((response) => {
+      return response.json();
+    })
+    .then((response) => {
+      return response.responses.map((response) => {
+        return response.responses.map((page) => page.fullTextAnnotation.text);
+      });
+    });
+  }
+
+  /**
+   * Extract text from Google Cloud Storage file.
+   *
+   * @param {string} gscloudfile URL from Google Cloud Storage file
+   * @returns {array} `[...[...string]]` [Documents[...Page{string}]]
+   */
+  analyzeFileFromCloudStorage(gscloudfile) {
     assert("The gscloud file address can't be blank.", !isBlank(gscloudfile));
 
     let accessToken = this.get('session.data.authenticated.access_token');
@@ -42,7 +94,7 @@ class GoogleVisionService extends Service {
       ]
     };
 
-    return fetch('https://content-vision.googleapis.com/v1/images:annotate?alt=json', {
+    return fetch('https://cors-anywhere.herokuapp.com/https://content-vision.googleapis.com/v1/images:annotate?alt=json', {
       headers: {
         accept: '*/*',
         authorization: `Bearer ${accessToken}`,
@@ -51,43 +103,12 @@ class GoogleVisionService extends Service {
       body: JSON.stringify(body),
       method: 'POST',
       credentials: 'include'
-      // referrerPolicy: 'no-referrer-when-downgrade'
     })
     .then((response) => {
       return response.json();
     })
     .then((response) => {
-      console.log(response.responses.firstObject.fullTextAnnotation.text);
-      return response.responses.firstObject.fullTextAnnotation.text;
+      return response.responses.map((document) => document.fullTextAnnotation.text);
     });
   }
 }
-
-export default GoogleVisionService;
-
-// fetch('https://content-vision.googleapis.com/v1/images:annotate?alt=json', {
-//   credentials: 'include',
-//   headers: {
-//     accept: '*/*',
-//     'accept-language': 'en-US,en;q=0.9,es-419;q=0.8,es;q=0.7',
-//     authorization:
-//       'Bearer ya29.ImW1BwgxLZpcavM9_jaQTqD9pnK6A4FUOIGyZ-mu0_SylS8i3xWco47B3Tayop8JTuo7HCDWEUcJVUigU9EUkNpe_GqqTW-HMxnIyBv61FjMmHQO4ApQSKejK8iWdaGjXl52fChhqw',
-//     'content-type': 'application/json',
-//     'sec-fetch-mode': 'cors',
-//     'sec-fetch-site': 'same-origin',
-//     'x-clientdetails':
-//       'appVersion=5.0%20(Windows%20NT%2010.0%3B%20Win64%3B%20x64)%20AppleWebKit%2F537.36%20(KHTML%2C%20like%20Gecko)%20Chrome%2F79.0.3945.79%20Safari%2F537.36&platform=Win32&userAgent=Mozilla%2F5.0%20(Windows%20NT%2010.0%3B%20Win64%3B%20x64)%20AppleWebKit%2F537.36%20(KHTML%2C%20like%20Gecko)%20Chrome%2F79.0.3945.79%20Safari%2F537.36',
-//     'x-goog-encode-response-if-executable': 'base64',
-//     'x-javascript-user-agent': 'apix/3.0.0 google-api-javascript-client/1.1.0',
-//     'x-origin': 'https://explorer.apis.google.com',
-//     'x-referer': 'https://explorer.apis.google.com',
-//     'x-requested-with': 'XMLHttpRequest'
-//   },
-//   referrer:
-//     'https://content-vision.googleapis.com/static/proxy.html?usegapi=1&jsh=m%3B%2F_%2Fscs%2Fapps-static%2F_%2Fjs%2Fk%3Doz.gapi.es.xhypXBFw-YI.O%2Fam%3DwQE%2Fd%3D1%2Fct%3Dzgms%2Frs%3DAGLTcCNIYS8P8MA2BPqaol3qYCZ18OvnqA%2Fm%3D__features__',
-//   referrerPolicy: 'no-referrer-when-downgrade',
-//   mode: 'cors'
-//   body:
-//     '{"requests":[{"features":[{"type":"LABEL_DETECTION"}],"image":{"source":{"imageUri":"gs://cloud-samples-data/vision/demo-img.jpg"}}}]}',
-//   method: 'POST',
-// });