+++ tests/unit/routes/page-not-found-test.js
@@ -1,0 +1,12 @@
+import { expect } from 'chai';
+import { describe, it } from 'mocha';
+import { setupTest } from 'ember-mocha';
+
+describe('Unit | Route | page-not-found', function() {
+  setupTest();
+
+  it('exists', function() {
+    let route = this.owner.lookup('route:page-not-found');
+    expect(route).to.be.ok;
+  });
+});