+++ CHANGELOG.md
@@ -5,7 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.0]
 ### Added
+- Added `jsconfig.json`.
+- Added `.prettierrc`
+- Added page not found (WIP).
+- Added OAuth with Google.
+- Added ember-file-upload, ember-simple-auth, ember-auto-import and torii addons.
+- Added service GoogleVision with methods `analyzeAttachFile` and `analyzeFileFromCloudStorage`.
 - Added @ember/edition-utils@^1.1.1 package.
 - Added @glimmer/component@^1.0.0-beta.1 package.
 - Added `setEdition("octane");` in `.ember-cli.js`.